fixup: モデルセレクター
